requirements.txt
# ---- Core -------------------------------------------------------------------
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1

# ---- LLM & Embeddings -------------------------------------------------------
google-genai>=1.0.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2

# ---- Vector store -----------------------------------------------------------
faiss-cpu>=1.8.0
numpy>=1.26.0

# ---- Document parsing -------------------------------------------------------
pymupdf>=1.24.0          # fitz — PDF text + image extraction
pdfplumber>=0.11.0       # table extraction
pytesseract>=0.3.13      # OCR wrapper
Pillow>=10.3.0           # image processing

# ---- NLP / Text splitting ---------------------------------------------------
langchain>=0.2.0
langchain-community>=0.2.0

# ---- UI ---------------------------------------------------------------------
streamlit>=1.35.0

# ---- Testing ----------------------------------------------------------------
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0            # FastAPI test client

# ---- Dev / Quality ----------------------------------------------------------
ruff>=0.4.0
mypy>=1.10.0
